Initial MinGW support
This commit adds initial support to Player for building with the MinGW compiler
suite on Windows or other platforms (via cross-compilers).  The support is 
initially limited to the core Player libraries, console-based utilities, and
drivers that don't require third-party dependencies.  Further support (gtk, 
boost, etc) is ongoing.

Currently this support is being tested using the MinGW compiler suite on
Windows 7, and using the MinGW 32 and 64 cross-compilers on Fedora 18, with
binaries run via WINE.