Fix off-by-one errors in ranger boundary checking
When accessing an array in the RangerProxy class (e.g. the ranges array), a
check is made to see if the number of elements in the array is grater than
the index being requested.  If the array had 0 elements (was unallocated), and
a request was made for the 0th element, this test passed and an unallocated
array was accessed.

This commit fixes the off-by-one error; you can only ever get index (length-1)
in any of the RangerProxy's array accesses.