Add support for hokuyoaist >= 3.0.0
The hokuyoaist driver was split out of the gearbox library
after version 9.11 and is now hosted on github.  The new
hokuyoaist library, version 3.0, removed support for setting
the baudrate of the sensor during runtime.  The driver has
been updated to reflect this change.  There was also an error
in detecting the version of the hokuyoaist library: the driver
was looking for version 3.0.0 exactly.  This has been fixed
so it looks for versions >=3.0.0.