Added target to run cppcheck
Added a target that will run cppcheck over the source code.  It uses
the cppcheck --project argument introduced in cppcheck 1.76, which
reads a json file of compiler commands that CMake generates with the
CMAKE_EXPORT_COMPILE_COMMANDS option.

The cppcheck output can be piped to file with a command like:
make cppcheck 2> cppcheck.log