Fix build errors with gcc7
This commit fixes two build errors exposed by gcc7 and Fedora's default
compile warnings.  The first error is a string format error in the ruby
bindings, where the "%s" argument must be supplied as a format string to
print an arbitrary string.

The second error is in the roboteq driver, where a pointer to a boolean
flag wasn't being properly dereferenced to determine if a motor was
enabled.  The comparision was looking at the pointer itself, which would
always test to true.